Whitelist repo/cache contents in .gitignore
The IP cache can be shared across computers, meaning that the first
build after a clone will be much faster if we version control the
cache contents. This will also grow the size of the project repository,
so we may want to weigh if this is worth it or not in the future.